refactor(miner): isolate internal Equihash solver
